new version

+++ src/app/earn/earn.page.ts
@@ -1,5 +1,9 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
-import { AlertController, IonInfiniteScroll, PopoverController } from '@ionic/angular';
+import {
+  AlertController,
+  IonInfiniteScroll,
+  PopoverController,
+} from '@ionic/angular';
 import { EarnService } from 'src/service/earn/earn.service';
 import { LoadingService } from 'src/service/loading/loading.service';
 import { PopIncomeComponent } from '../component/pop-income/pop-income.component';
@@ -27,7 +31,7 @@ export class EarnPage implements OnInit {
     private loadSerivce: LoadingService,
     private alertController: AlertController,
     public popoverController: PopoverController
-    ) {}
+  ) {}
 
   ngOnInit() {
     this.loadEarn();
@@ -47,37 +51,14 @@ export class EarnPage implements OnInit {
     });
   }
 
-  onAdd() {
-    if (!this.amount_money) {
-      this.msg = 'ກະລຸນາປ້ອນຈຳນວນເງິນ!';
-      return;
-    } else if (this.amount_money % 1000 != 0 || this.amount_money <= 0) {
-      this.msg = 'ຈຳນວນເງິນບໍຖືກຕ້ອງ!';
-      return;
-    } else {
-      this.msg = '';
-    }
-
-    this.loadSerivce.onLoading();
-    this.earn
-      .addEarn(this.token, { money: this.amount_money })
-      .subscribe((res: any) => {
-        this.loadSerivce.onDismiss();
-
-        console.log('earn data=>', res);
-        this.amount_money = null;
-
-        this.loadEarn();
-      });
-    this.popEarn = false;
-  }
-
-  onDel(earn_id: string, i: number) {
+  async onDel(earn_id: string, i: number) {
     console.log('id=>', earn_id, i);
-    this.loadSerivce.onLoading();
-    this.earn.deleteEarn(this.token, earn_id).subscribe((res: any) => {
-      this.loadSerivce.onDismiss();
 
+    const confirm = await this.loadSerivce.alertConfirm('ຕ້ອງການລຶບອອກແທ້ບໍ?');
+
+    if (!confirm) return;
+
+    this.earn.deleteEarn(this.token, earn_id).subscribe((res: any) => {
       console.log('delete earn=>', res);
       if (res.success) {
         this.loadEarn();
@@ -86,6 +67,7 @@ export class EarnPage implements OnInit {
   }
   doRefresh(event) {
     setTimeout(() => {
+      this.infiniteScroll.disabled = false;
       this.loadEarn();
       event.target.complete();
     }, 1000);
@@ -94,17 +76,10 @@ export class EarnPage implements OnInit {
   loadMore(e: any) {
     this.earn.getAllEarn(this.token, this.skip).subscribe(async (res: any) => {
       this.loadSerivce.onDismiss();
-
       console.log('all data=>', this.skip, res);
       if (res.success) {
-        if (res.data.earn.length == 0) {
-          const alert = await this.alertController.create({
-            cssClass: '_div',
-            header: 'ແຈ້ງເຕືອນ',
-            message: 'ບໍ່ມີຂໍ້ມູນເພີ່ມເຕີມ!!!',
-            buttons: ['ປິດ'],
-          });
-          await alert.present();
+        if (!res.data.earn.length) {
+          this.infiniteScroll.disabled = true;
           return;
         }
         this.earnData = this.earnData.concat(res.data.earn);
@@ -125,12 +100,12 @@ export class EarnPage implements OnInit {
     this.infiniteScroll.disabled = !this.infiniteScroll.disabled;
   }
 
-  async openPopIncome(ev:any){
+  async openPopIncome(ev: any) {
     const popover = await this.popoverController.create({
       component: PopIncomeComponent,
       cssClass: 'pop-over-style',
       componentProps: {
-        action:'earn'
+        action: 'earn',
       },
       event: ev,
       translucent: true,
@@ -138,12 +113,12 @@ export class EarnPage implements OnInit {
     popover.style.cssText = '--width: 94%';
 
     await popover.present();
-  
+
     const { role } = await popover.onDidDismiss();
     console.log('onDidDismiss resolved with role', role);
-    
-    if (role == 'success') {      
-      this.loadEarn()
+
+    if (role == 'success') {
+      this.loadEarn();
     }
   }
 }
